requirements.txt
Flask>=2.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
Flask-Cors>=4.0.0